+++ MatomoTracker.php
@@ -223,7 +223,8 @@ class MatomoTracker
             !empty($_SERVER['HTTP_SEC_CH_UA_PLATFORM']) ? $_SERVER['HTTP_SEC_CH_UA_PLATFORM'] : '',
             !empty($_SERVER['HTTP_SEC_CH_UA_PLATFORM_VERSION']) ? $_SERVER['HTTP_SEC_CH_UA_PLATFORM_VERSION'] : '',
             !empty($_SERVER['HTTP_SEC_CH_UA_FULL_VERSION_LIST']) ? $_SERVER['HTTP_SEC_CH_UA_FULL_VERSION_LIST'] : '',
-            !empty($_SERVER['HTTP_SEC_CH_UA_FULL_VERSION']) ? $_SERVER['HTTP_SEC_CH_UA_FULL_VERSION'] : ''
+            !empty($_SERVER['HTTP_SEC_CH_UA_FULL_VERSION']) ? $_SERVER['HTTP_SEC_CH_UA_FULL_VERSION'] : '',
+            !empty($_SERVER['HTTP_SEC_CH_UA_FORM_FACTORS']) ? $_SERVER['HTTP_SEC_CH_UA_FORM_FACTORS'] : ''
         );
         if (!empty($apiUrl)) {
             self::$URL = $apiUrl;
@@ -593,6 +594,8 @@ class MatomoTracker
      *      or an array containing all brands with the structure
      *      [['brand' => 'Chrome', 'version' => '10.0.2'], ['brand' => '...]
      * @param string $uaFullVersion  Value of the header 'HTTP_SEC_CH_UA_FULL_VERSION'
+     * @param string|array<string> $formFactors  Value of the header 'HTTP_SEC_CH_UA_FORM_FACTORS'
+     *      or an array containing all form factors with structure ["Desktop", "XR"]
      *
      * @return $this
      */
@@ -601,7 +604,8 @@ class MatomoTracker
         string $platform = '',
         string $platformVersion = '',
         $fullVersionList = '',
-        string $uaFullVersion = ''
+        string $uaFullVersion = '',
+        $formFactors = ''
     ) {
         if (is_string($fullVersionList)) {
             $reg  = '/^"([^"]+)"; ?v="([^"]+)"(?:, )?/';
@@ -617,12 +621,25 @@ class MatomoTracker
             $fullVersionList = [];
         }
 
+        if (is_string($formFactors)) {
+            $formFactors = explode(',', $formFactors);
+            $formFactors = array_filter(array_map(
+                function ($item) {
+                    return trim($item, '" ');
+                },
+                $formFactors
+            ));
+        } elseif (!is_array($formFactors)) {
+            $formFactors = [];
+        }
+
         $this->clientHints = array_filter([
             'model' => $model,
             'platform' => $platform,
             'platformVersion' => $platformVersion,
             'uaFullVersion' => $uaFullVersion,
             'fullVersionList' => $fullVersionList,
+            'formFactors' => $formFactors,
         ]);
 
         return $this;
@@ -810,7 +827,7 @@ class MatomoTracker
 
         return $this->sendRequest($url);
     }
-			 
+
     /**
      * Override PageView id for every use of `doTrackPageView()`. Do not use this if you call `doTrackPageView()`
      * multiple times during tracking (if, for example, you are tracking a single page application).
@@ -2097,7 +2114,7 @@ didn't change any existing VisitorId value */
             curl_setopt_array($ch, $options);
             ob_start();
             $response = @curl_exec($ch);
-            
+
             try {
                 $header = '';