+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 This is the Developer Changelog for Matomo PHP Tracker. All breaking changes or new features are listed below.
 
-## Matomo PHP Tracker 3.4.0
+## Matomo PHP Tracker 4.0.0
 ### Changed
 - a lot of arguments of `MatomoTracker` methods have explicitly types
 - a lot of `MatomoTracker` method return types have strict types
+
+## Matomo PHP Tracker 3.3.2
+### Changed
+- Support for formFactors client hint parameter, supported as of Matomo 5.2.0
 
 ## Matomo PHP Tracker 3.3.1
 ### Fixed